Adding safe I/O code => NO MORE JELLY BEAN!
Removed code to use the clipboard for import/export (nobody liked it anyway)
Added code to use system file picker for import/export :)
Added initial support for rendering tracks to file. I personally don't think this is any useful, but others seem to think otherwise...

File access is done via ACTION_OPEN_DOCUMENT / ACTION_CREATE_DOCUMENT which are API 19. So minsdk is now 19 (was 17) which means KitKat (October 2013) and newer :(
Also, a lot of pointless formatting changes in an attempt to slowly change all files to a unified style.